% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTable.R
\name{LTable}
\alias{LTable}
\title{Make LaTeX table code from R objcts.}
\usage{
LTable(..., caption = "", label = "", indent = F, file = "")
}
\arguments{
\item{...}{can be matrices, data.frames, or lists of matrices or data.frames.}

\item{caption}{a character string of the caption.}

\item{label}{a character string of the label.}

\item{indent}{a boolean indicating if indent.}

\item{file}{a connection, or a character string naming the file to print to. If "" (the default), ltable prints to the standard output connection, the console unless redirected by sink. If it is "|cmd", the output is piped to the command given by 'cmd', by opening a pipe connection.}
}
\description{
This function makes LaTeX table code from R objcts.
}
\details{
The function saves LaTeX users' time for writing tables. It uses directly the results from R to produce LaTeX table code.

The function handels values in \code{...} in the following way:

Elements in \code{...} will be binded by row.

If an element in \code{...} is a list, then its element will be binded by column.

Missing values will be filled by empty character "".
}
\examples{
x = matrix(rnorm(9),3,3)
y = data.frame(a=c('a','b','c','d'),b=1:4)
z = matrix(0,2,5)
LTable(y,caption="Hello!",list(round(x,2),z),label="tab:hi",indent=TRUE)

}
\author{
Yukai Yang, \email{yukai.yang@statistik.uu.se}
}
\keyword{LTable}
