% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ltable.R
\name{plot_ltable}
\alias{plot_ltable}
\title{plot ltable as tree}
\usage{
plot_ltable(
  ltable,
  plot_hybrid = FALSE,
  use_factor_ID = FALSE,
  modify_theme = TRUE
)
}
\arguments{
\item{ltable}{ltable}
}
\description{
plot ltable as tree
}
