\name{sdata}
\alias{sdata}
\docType{data}
\title{
Simple Data.
}
\description{
This data has no other meaning and purpose except for package functionality presentation.
}
\usage{data(sdata)}
\format{
  A data frame with 22 observations (some values are purposely missing) on the following 4 variables.
  \describe{
    \item{\code{a}}{a logical vector}
    \item{\code{b}}{a numeric vector}
    \item{\code{c}}{a factor with levels \code{female} and \code{male}}
    \item{\code{d}}{a character vector with variants "A" and "B"}
  }
}
\details{
You can construct tables with data set fields of factor, character, logical, and numeric classes.
}


\examples{
data(sdata, package="ltable")
}
\keyword{datasets}
