% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bd.R
\name{sim_bd}
\alias{sim_bd}
\title{Function to simulate a tree, conditional on observing all states.}
\usage{
sim_bd(
  parameters,
  crown_age,
  max_spec = 1e+05,
  min_spec = 2,
  max_species_extant = TRUE,
  tree_size_hist = FALSE,
  non_extinction = TRUE,
  verbose = FALSE,
  max_tries = 1e+06,
  drop_extinct = FALSE,
  start_at_crown = TRUE,
  seed = NULL
)
}
\arguments{
\item{parameters}{vector with two parameters consisting of per species rates
of: speciating, extinction.}

\item{crown_age}{age of the crown}

\item{max_spec}{condition on tree not being larger than this}

\item{min_spec}{condition on tree not being smaller than this}

\item{max_species_extant}{should the maximum number of species only include
extant species (if TRUE), or also extinct species?}

\item{tree_size_hist}{returns a histogram of tree sizes, including failed
trees (e.g. trees that went extinct or were outside conditioning criteria)}

\item{non_extinction}{condition on non-extinction if TRUE}

\item{verbose}{provide verbose output}

\item{max_tries}{maximum number of tries before giving up}

\item{drop_extinct}{should extinct lineages be returned?}

\item{start_at_crown}{should the simulation start at the root, or the crown?}

\item{seed}{random seed
By default, the algorithm keeps simulating until it generates a tree where
both crown lineages survive to the present - this is to ensure that the tree
has a crown age that matches the used crown age. You can modify
'non-extinction' to deviate from this behaviour.}
}
\description{
Function to simulate a tree, conditional on observing all states.
}
